Fix signature for glfwSetMonitorUserPointer
Fiddle importer is raising an error that `const` is not a type when parsing the signature of `glfwSetMonitorUserPointer`. Removing the `const` fixes this. Based on GLFW's header files, this is correct since it does not have a `const` in the signature: https://github.com/glfw/glfw/blob/814b7929c5add4b0541ccad26fb81f28b71dc4d8/include/GLFW/glfw3.h#L2222